Webview async resolution, take 2 (#43)
* First stab at lifecycling buses for lazy resolution

* No longer compiles with generic map interface

* Moved method into respective section

* Broadcast last known state to newly connecting clients

* Lets have sourceMaps

* Make disposal optional

* Allow for async resolution in all sandboxes

* Use special operator for resolution debouncing to optimize for minimal latency

* Push complexity of channel impl down into base class

* Deps in sync

* Bring this back

* Simplification

* Figured out testing. however, tests need refactoring.

* Drop it slightly for now

* Sort imports

* This should only be available on the bus

* Notify all clients whenever a new one connects

* Cast not necessary

* Pin rxjs

* Revert "Pin rxjs"

This reverts commit 97a37a7bb6d09038b3ac088529502eb9419f136d.

* remove dist dir

* avoid use any if possible

* update deps

* use ts file for rollup configuration

* correct test coverage treshold

* Release 2.2.0-beta.0

* allow falsy event values

* make worker example more clear

* minor code formatting

* don't emit if state hasn't changed

* Release 3.0.0-beta.0

* no sourcemaps for published code

* fetch depth 0

* consolidate release pipeline

* Release 3.0.0-beta.1

* use distinctUntilChanged

* pass in subscriber

+++ examples/worker_threads.js
@@ -4,7 +4,9 @@ import Channel from '../dist/esm/worker_threads.js';
 
 const filename = new URL('', import.meta.url).pathname;
 
-const ch = new Channel('test', {});
+const ch = new Channel('test', {
+    shutdownWorker: false
+});
 
 if (isMainThread) {
     const bus = await ch.registerPromise([
@@ -14,9 +16,13 @@ if (isMainThread) {
     ]);
 
     bus.listen('onCustomEvent', (msg) =>
-        console.log('Received from worker thread:', msg));
+        msg && console.log('Received from worker thread:', msg));
 
-    bus.listen('onExit', () => {
+    bus.listen('shutdownWorker', (shutdownWorker) => {
+        if (!shutdownWorker) {
+            return;
+        }
+
         console.log('Bye bye');
         ch.providers.map((p) => p.terminate());
     });
@@ -41,7 +47,7 @@ if (isMainThread) {
      */
     client.listen('onCustomWorkerEvent', (id) => {
         if (workerData.id === id) {
-            client.broadcast({ 'onExit': workerData.id });
+            client.broadcast({ shutdownWorker: true });
         }
     });
 }
